Add Koch Curve, Snowflake, Hexagon, Square and Towers of Hanoi
I probably went a bit overboard with all these shapes. I liked the Koch
Snowflake when I saw it on Wikipedia.

My initial attempt was to recreate exercises I had to do for university
ages ago. This turned out to be pretty easy with the better
understanding for recursion that I have now.